Allow runtime parameters to be passed as a pack
This way we can keep the uniform function call API and just let the user
override what they need to for their particular search (this will need to be
documented since it seems tricky to get into the type system for now).

It could be possible to use a dispatcher function that checks the type of the
search and gives the API based on this and we fill in the defaults behind the
scenes, but that seems overkill.