Remove macros by make Incumbent component untyped
Instead we type the functions we call on the incumbent. Although the call sites
look a bit strange, it lets us have a lot more control over the inner component
types and to avoid the user needing to call the correct macro with the correct
types just to get things working.